Fixed Employee

+++ src/main/java/com/bootcamp/miniproject/service/EmployeeService.java
@@ -65,6 +65,7 @@ public class EmployeeService {
 
 //UPDATE	
 	public void update(Employee employee) {
+		
 		Employee emp = new Employee();
 		emp.setId(employee.getId());
 		emp.setFirstName(employee.getFirstName());
@@ -74,7 +75,7 @@ public class EmployeeService {
 		emp.setHaveAccount(employee.isHaveAccount());
 		emp.setActive(employee.isActive());
 		emp.setCreatedBy(employee.getCreatedBy());
-		//emp.setModifiedBy(employee.getModifiedBy());
+		emp.setModifiedBy(employee.getModifiedBy());
 		employeeDao.update(emp);
 		
 		
@@ -94,25 +95,40 @@ public class EmployeeService {
 			}
 		}
 		System.out.println("haveAccount: "+ employee.isHaveAccount());
-		if(employee.isHaveAccount()) {
-			System.out.println("gak didelete");
-			User user = new User();
-			user.setId(employee.getUser().getId());
-			user.setEmployee(emp);
-			user.setActive(employee.getUser().isActive());
-			user.setRole(employee.getUser().getRole());
-			user.setUsername(employee.getUser().getUsername());
-			user.setPassword(employee.getUser().getPassword());
-			user.setCreatedBy(employee.getUser().getCreatedBy());
-			//user.setModifiedBy(employee.getUser().getModifiedBy());
-			userDao.saveOrUpdate(user);
-		}else {
-			System.out.println("delete cuy");
-			User user = new User();
-			user.setId(employee.getUser().getId());
-			userDao.delete(user);
+		if (employee.getUser().getId()!=0) {
+			if(employee.isHaveAccount()) {
+				System.out.println("gak didelete");
+				User user = new User();
+				user.setId(employee.getUser().getId());
+				user.setEmployee(emp);
+				user.setActive(employee.getUser().isActive());
+				user.setRole(employee.getUser().getRole());
+				user.setUsername(employee.getUser().getUsername());
+				user.setPassword(employee.getUser().getPassword());
+				user.setCreatedBy(employee.getUser().getCreatedBy());
+				user.setModifiedBy(employee.getUser().getModifiedBy());
+				userDao.saveOrUpdate(user);
+			}else {
+				System.out.println("delete cuy");
+				User user = new User();
+				user.setId(employee.getUser().getId());
+				userDao.delete(user);
 
+			}
+		} else {
+			if(employee.isHaveAccount()) {
+				User user = new User();
+				user.setEmployee(emp);
+				user.setActive(employee.getUser().isActive());
+				user.setRole(employee.getUser().getRole());
+				user.setUsername(employee.getUser().getUsername());
+				user.setPassword(employee.getUser().getPassword());
+				user.setCreatedBy(employee.getUser().getCreatedBy());
+				user.setModifiedBy(employee.getUser().getModifiedBy());
+				userDao.save(user);
+			}
 		}
+		
 	}
